refactor: v0.12.2 — split chart.rs into 6 modules
