+++ src/chart.rs
@@ -5,7 +5,23 @@
 //! [`Context::histogram`](crate::Context::histogram) directly.
 
 use crate::style::{Color, Style};
-use unicode_width::UnicodeWidthStr;
+
+mod axis;
+mod bar;
+mod braille;
+mod grid;
+mod render;
+
+pub(crate) use bar::build_histogram_config;
+pub(crate) use render::render_chart;
+
+use axis::{build_tui_ticks, format_number, resolve_bounds, TickSpec};
+use bar::draw_bar_dataset;
+use braille::draw_braille_dataset;
+use grid::{
+    apply_grid, build_legend_items, build_x_tick_col_map, build_y_tick_row_map, center_text,
+    map_value_to_cell, marker_char, overlay_legend_on_plot, sturges_bin_count, GridSpec,
+};
 
 const BRAILLE_BASE: u32 = 0x2800;
 const BRAILLE_LEFT_BITS: [u32; 4] = [0x01, 0x02, 0x04, 0x40];
@@ -501,7 +517,7 @@ impl ChartRenderer {
                 let mut cursor = 0usize;
 
                 for (segment, style) in row.segments {
-                    let width = UnicodeWidthStr::width(segment.as_str());
+                    let width = unicode_width::UnicodeWidthStr::width(segment.as_str());
                     line.push_str(&segment);
                     if let Some(color) = style.fg {
                         spans.push((cursor, cursor + width, color));
@@ -513,1227 +529,4 @@ impl ChartRenderer {
             })
             .collect()
     }
-}
-
-/// Build a histogram chart configuration from raw values.
-pub(crate) fn build_histogram_config(
-    data: &[f64],
-    options: &HistogramBuilder,
-    width: u32,
-    height: u32,
-    axis_style: Style,
-) -> ChartConfig {
-    let mut sorted: Vec<f64> = data.iter().copied().filter(|v| v.is_finite()).collect();
-    sorted.sort_by(f64::total_cmp);
-
-    if sorted.is_empty() {
-        return ChartConfig {
-            title: Some("Histogram".to_string()),
-            title_style: None,
-            x_axis: Axis {
-                title: options.x_title.clone(),
-                bounds: Some((0.0, 1.0)),
-                labels: None,
-                ticks: None,
-                title_style: None,
-                style: axis_style,
-            },
-            y_axis: Axis {
-                title: options.y_title.clone(),
-                bounds: Some((0.0, 1.0)),
-                labels: None,
-                ticks: None,
-                title_style: None,
-                style: axis_style,
-            },
-            datasets: Vec::new(),
-            legend: LegendPosition::None,
-            grid: true,
-            grid_style: None,
-            hlines: Vec::new(),
-            vlines: Vec::new(),
-            frame_visible: true,
-            x_axis_visible: true,
-            y_axis_visible: true,
-            width,
-            height,
-        };
-    }
-
-    let n = sorted.len();
-    let min = sorted[0];
-    let max = sorted[n.saturating_sub(1)];
-    let bin_count = options.bins.unwrap_or_else(|| sturges_bin_count(n));
-
-    let span = if (max - min).abs() < f64::EPSILON {
-        1.0
-    } else {
-        max - min
-    };
-    let bin_width = span / bin_count as f64;
-
-    let mut counts = vec![0usize; bin_count];
-    for value in sorted {
-        let raw = ((value - min) / bin_width).floor();
-        let mut idx = if raw.is_finite() { raw as isize } else { 0 };
-        if idx < 0 {
-            idx = 0;
-        }
-        if idx as usize >= bin_count {
-            idx = (bin_count.saturating_sub(1)) as isize;
-        }
-        counts[idx as usize] = counts[idx as usize].saturating_add(1);
-    }
-
-    let mut data_points = Vec::with_capacity(bin_count);
-    for (i, count) in counts.iter().enumerate() {
-        let center = min + (i as f64 + 0.5) * bin_width;
-        data_points.push((center, *count as f64));
-    }
-
-    let mut labels: Vec<String> = Vec::new();
-    let step = (bin_count / 4).max(1);
-    for i in (0..=bin_count).step_by(step) {
-        let edge = min + i as f64 * bin_width;
-        labels.push(format_number(edge, bin_width));
-    }
-
-    ChartConfig {
-        title: Some("Histogram".to_string()),
-        title_style: None,
-        x_axis: Axis {
-            title: options.x_title.clone(),
-            bounds: Some((min, max.max(min + bin_width))),
-            labels: Some(labels),
-            ticks: None,
-            title_style: None,
-            style: axis_style,
-        },
-        y_axis: Axis {
-            title: options.y_title.clone(),
-            bounds: Some((0.0, counts.iter().copied().max().unwrap_or(1) as f64)),
-            labels: None,
-            ticks: None,
-            title_style: None,
-            style: axis_style,
-        },
-        datasets: vec![Dataset {
-            name: "Histogram".to_string(),
-            data: data_points,
-            color: options.color,
-            marker: Marker::Block,
-            graph_type: GraphType::Bar,
-            up_color: None,
-            down_color: None,
-        }],
-        legend: LegendPosition::None,
-        grid: true,
-        grid_style: None,
-        hlines: Vec::new(),
-        vlines: Vec::new(),
-        frame_visible: true,
-        x_axis_visible: true,
-        y_axis_visible: true,
-        width,
-        height,
-    }
-}
-
-/// Render a chart into styled row segments.
-pub(crate) fn render_chart(config: &ChartConfig) -> Vec<ChartRow> {
-    let width = config.width as usize;
-    let height = config.height as usize;
-    if width == 0 || height == 0 {
-        return Vec::new();
-    }
-
-    let frame_style = config.x_axis.style;
-    let dim_style = Style::new().dim();
-    let axis_style = config.y_axis.style;
-    let title_style = Style::new()
-        .bold()
-        .fg(config.x_axis.style.fg.unwrap_or(Color::White));
-    let title_style = config.title_style.unwrap_or(title_style);
-
-    let title_rows = usize::from(config.title.is_some());
-    let has_x_title = config.x_axis_visible && config.x_axis.title.is_some();
-    let x_title_rows = usize::from(has_x_title);
-    let frame_rows = if config.frame_visible { 2 } else { 0 };
-    let x_axis_rows = if config.x_axis_visible {
-        2 + x_title_rows
-    } else {
-        0
-    };
-
-    // Row budget: title + top_frame + plot + axis_line + x_labels + [x_title] + bottom_frame
-    //           = title_rows + 1 + plot_height + 1 + 1 + x_title_rows + 1
-    //           = title_rows + plot_height + 3 + x_title_rows
-    // Solve for plot_height:
-    let overhead = title_rows + frame_rows + x_axis_rows;
-    if height <= overhead || width < 3 {
-        return minimal_chart(config, width, frame_style, title_style);
-    }
-    let plot_height = height.saturating_sub(overhead).max(1);
-
-    let (x_min, x_max) = resolve_bounds(
-        config
-            .datasets
-            .iter()
-            .flat_map(|d| d.data.iter().map(|p| p.0)),
-        config.x_axis.bounds,
-    );
-    let (y_min, y_max) = resolve_bounds(
-        config
-            .datasets
-            .iter()
-            .flat_map(|d| d.data.iter().map(|p| p.1)),
-        config.y_axis.bounds,
-    );
-
-    let y_label_chars: Vec<char> = if config.y_axis_visible {
-        config
-            .y_axis
-            .title
-            .as_deref()
-            .map(|t| t.chars().collect())
-            .unwrap_or_default()
-    } else {
-        Vec::new()
-    };
-    let y_label_col_width = if y_label_chars.is_empty() { 0 } else { 2 };
-
-    let legend_items = build_legend_items(&config.datasets);
-    let legend_on_right = matches!(
-        config.legend,
-        LegendPosition::TopRight | LegendPosition::BottomRight
-    );
-    let legend_width = if legend_on_right && !legend_items.is_empty() {
-        legend_items
-            .iter()
-            .map(|(_, name, _)| 4 + UnicodeWidthStr::width(name.as_str()))
-            .max()
-            .unwrap_or(0)
-    } else {
-        0
-    };
-
-    let y_ticks = if let Some(ref manual) = config.y_axis.ticks {
-        TickSpec {
-            values: manual.clone(),
-            step: if manual.len() > 1 {
-                manual[1] - manual[0]
-            } else {
-                1.0
-            },
-        }
-    } else {
-        build_tui_ticks(y_min, y_max, plot_height)
-    };
-    let y_min = y_ticks.values.first().copied().unwrap_or(y_min).min(y_min);
-    let y_max = y_ticks.values.last().copied().unwrap_or(y_max).max(y_max);
-
-    let use_manual_y_labels = config.y_axis.ticks.is_some() && config.y_axis.labels.is_some();
-    let y_tick_labels: Vec<String> = if use_manual_y_labels {
-        config
-            .y_axis
-            .labels
-            .as_deref()
-            .unwrap_or(&[])
-            .iter()
-            .take(y_ticks.values.len())
-            .cloned()
-            .collect()
-    } else {
-        y_ticks
-            .values
-            .iter()
-            .map(|v| format_number(*v, y_ticks.step))
-            .collect()
-    };
-    let y_tick_width = y_tick_labels
-        .iter()
-        .map(|s| UnicodeWidthStr::width(s.as_str()))
-        .max()
-        .unwrap_or(1);
-    let y_axis_width = if config.y_axis_visible {
-        y_tick_width + 2
-    } else {
-        0
-    };
-
-    let inner_width = if config.frame_visible {
-        width.saturating_sub(2)
-    } else {
-        width
-    };
-    let plot_width = inner_width
-        .saturating_sub(y_label_col_width)
-        .saturating_sub(y_axis_width)
-        .saturating_sub(legend_width)
-        .max(1);
-    let content_width = y_label_col_width + y_axis_width + plot_width + legend_width;
-
-    let x_ticks = if let Some(ref manual) = config.x_axis.ticks {
-        TickSpec {
-            values: manual.clone(),
-            step: if manual.len() > 1 {
-                manual[1] - manual[0]
-            } else {
-                1.0
-            },
-        }
-    } else {
-        build_tui_ticks(x_min, x_max, plot_width)
-    };
-    let x_min = x_ticks.values.first().copied().unwrap_or(x_min).min(x_min);
-    let x_max = x_ticks.values.last().copied().unwrap_or(x_max).max(x_max);
-
-    let mut plot_chars = vec![vec![' '; plot_width]; plot_height];
-    let mut plot_styles = vec![vec![Style::new(); plot_width]; plot_height];
-
-    apply_grid(
-        config,
-        GridSpec {
-            x_ticks: &x_ticks.values,
-            y_ticks: &y_ticks.values,
-            x_min,
-            x_max,
-            y_min,
-            y_max,
-        },
-        &mut plot_chars,
-        &mut plot_styles,
-        config.grid_style.unwrap_or(dim_style),
-    );
-
-    for &(y_val, ref style) in &config.hlines {
-        let row = map_value_to_cell(y_val, y_min, y_max, plot_height, true);
-        if row < plot_height {
-            for col in 0..plot_width {
-                plot_chars[row][col] = '─';
-                plot_styles[row][col] = *style;
-            }
-        }
-    }
-    for &(x_val, ref style) in &config.vlines {
-        let col = map_value_to_cell(x_val, x_min, x_max, plot_width, false);
-        if col < plot_width {
-            for row in 0..plot_height {
-                if plot_chars[row][col] == ' ' || plot_chars[row][col] == '·' {
-                    plot_chars[row][col] = '│';
-                    plot_styles[row][col] = *style;
-                }
-            }
-        }
-    }
-
-    for dataset in &config.datasets {
-        match dataset.graph_type {
-            GraphType::Line | GraphType::Area | GraphType::Scatter => {
-                draw_braille_dataset(
-                    dataset,
-                    x_min,
-                    x_max,
-                    y_min,
-                    y_max,
-                    &mut plot_chars,
-                    &mut plot_styles,
-                );
-            }
-            GraphType::Bar => {
-                draw_bar_dataset(
-                    dataset,
-                    x_min,
-                    x_max,
-                    y_min,
-                    y_max,
-                    &mut plot_chars,
-                    &mut plot_styles,
-                );
-            }
-        }
-    }
-
-    if !legend_items.is_empty()
-        && matches!(
-            config.legend,
-            LegendPosition::TopLeft | LegendPosition::BottomLeft
-        )
-    {
-        overlay_legend_on_plot(
-            config.legend,
-            &legend_items,
-            &mut plot_chars,
-            &mut plot_styles,
-            axis_style,
-        );
-    }
-
-    let y_tick_rows = build_y_tick_row_map(
-        &y_ticks.values,
-        if use_manual_y_labels {
-            config.y_axis.labels.as_deref()
-        } else {
-            None
-        },
-        y_min,
-        y_max,
-        plot_height,
-    );
-    let x_tick_cols = build_x_tick_col_map(
-        &x_ticks.values,
-        config.x_axis.labels.as_deref(),
-        config.x_axis.ticks.is_some() && config.x_axis.labels.is_some(),
-        x_min,
-        x_max,
-        plot_width,
-    );
-
-    let mut rows: Vec<ChartRow> = Vec::with_capacity(height);
-
-    // --- Title row ---
-    if let Some(title) = &config.title {
-        rows.push(ChartRow {
-            segments: vec![(center_text(title, width), title_style)],
-        });
-    }
-
-    if config.frame_visible {
-        rows.push(ChartRow {
-            segments: vec![(format!("┌{}┐", "─".repeat(content_width)), frame_style)],
-        });
-    }
-
-    let y_label_start = if y_label_chars.is_empty() {
-        0
-    } else {
-        plot_height.saturating_sub(y_label_chars.len()) / 2
-    };
-    let y_title_style = config.y_axis.title_style.unwrap_or(axis_style);
-
-    let zero_label = format_number(0.0, y_ticks.step);
-    for row in 0..plot_height {
-        let mut segments: Vec<(String, Style)> = Vec::new();
-        if config.frame_visible {
-            segments.push(("│".to_string(), frame_style));
-        }
-
-        if config.y_axis_visible {
-            if y_label_col_width > 0 {
-                let label_idx = row.wrapping_sub(y_label_start);
-                if label_idx < y_label_chars.len() {
-                    segments.push((format!("{} ", y_label_chars[label_idx]), y_title_style));
-                } else {
-                    segments.push(("  ".to_string(), Style::new()));
-                }
-            }
-
-            let (label, divider) =
-                if let Some(index) = y_tick_rows.iter().position(|(r, _)| *r == row) {
-                    let is_zero = y_tick_rows[index].1 == zero_label;
-                    (
-                        y_tick_rows[index].1.clone(),
-                        if is_zero { '┼' } else { '┤' },
-                    )
-                } else {
-                    (String::new(), '│')
-                };
-            let padded = format!("{:>w$}", label, w = y_tick_width);
-            segments.push((padded, axis_style));
-            segments.push((format!("{divider} "), axis_style));
-        }
-
-        let mut current_style = Style::new();
-        let mut buffer = String::new();
-        for col in 0..plot_width {
-            let style = plot_styles[row][col];
-            if col == 0 {
-                current_style = style;
-            }
-            if style != current_style {
-                if !buffer.is_empty() {
-                    segments.push((buffer.clone(), current_style));
-                    buffer.clear();
-                }
-                current_style = style;
-            }
-            buffer.push(plot_chars[row][col]);
-        }
-        if !buffer.is_empty() {
-            segments.push((buffer, current_style));
-        }
-
-        if legend_on_right && legend_width > 0 {
-            let legend_row = match config.legend {
-                LegendPosition::TopRight => row,
-                LegendPosition::BottomRight => {
-                    row.wrapping_add(legend_items.len().saturating_sub(plot_height))
-                }
-                _ => usize::MAX,
-            };
-            if let Some((symbol, name, color)) = legend_items.get(legend_row) {
-                let raw = format!("  {symbol} {name}");
-                let raw_w = UnicodeWidthStr::width(raw.as_str());
-                let pad = legend_width.saturating_sub(raw_w);
-                let text = format!("{raw}{}", " ".repeat(pad));
-                segments.push((text, Style::new().fg(*color)));
-            } else {
-                segments.push((" ".repeat(legend_width), Style::new()));
-            }
-        }
-
-        if config.frame_visible {
-            segments.push(("│".to_string(), frame_style));
-        }
-        rows.push(ChartRow { segments });
-    }
-
-    if config.x_axis_visible {
-        let mut axis_line = vec!['─'; plot_width];
-        for (col, _) in &x_tick_cols {
-            if *col < plot_width {
-                axis_line[*col] = '┬';
-            }
-        }
-        let footer_legend_pad = " ".repeat(legend_width);
-        let footer_ylabel_pad = if config.y_axis_visible {
-            " ".repeat(y_label_col_width)
-        } else {
-            String::new()
-        };
-
-        let mut axis_segments: Vec<(String, Style)> = Vec::new();
-        if config.frame_visible {
-            axis_segments.push(("│".to_string(), frame_style));
-        }
-        if config.y_axis_visible {
-            axis_segments.push((footer_ylabel_pad.clone(), Style::new()));
-            axis_segments.push((" ".repeat(y_tick_width), axis_style));
-            axis_segments.push(("┴─".to_string(), axis_style));
-        }
-        axis_segments.push((axis_line.into_iter().collect(), axis_style));
-        axis_segments.push((footer_legend_pad.clone(), Style::new()));
-        if config.frame_visible {
-            axis_segments.push(("│".to_string(), frame_style));
-        }
-        rows.push(ChartRow {
-            segments: axis_segments,
-        });
-
-        let mut x_label_line: Vec<char> = vec![' '; plot_width];
-        let mut occupied_until: usize = 0;
-        for (col, label) in &x_tick_cols {
-            if label.is_empty() {
-                continue;
-            }
-            let label_width = UnicodeWidthStr::width(label.as_str());
-            let start = col
-                .saturating_sub(label_width / 2)
-                .min(plot_width.saturating_sub(label_width));
-            if start < occupied_until {
-                continue;
-            }
-            for (offset, ch) in label.chars().enumerate() {
-                let idx = start + offset;
-                if idx < plot_width {
-                    x_label_line[idx] = ch;
-                }
-            }
-            occupied_until = start + label_width + 1;
-        }
-
-        let mut x_label_segments: Vec<(String, Style)> = Vec::new();
-        if config.frame_visible {
-            x_label_segments.push(("│".to_string(), frame_style));
-        }
-        if config.y_axis_visible {
-            x_label_segments.push((footer_ylabel_pad.clone(), Style::new()));
-            x_label_segments.push((" ".repeat(y_axis_width), Style::new()));
-        }
-        x_label_segments.push((x_label_line.into_iter().collect(), axis_style));
-        x_label_segments.push((footer_legend_pad.clone(), Style::new()));
-        if config.frame_visible {
-            x_label_segments.push(("│".to_string(), frame_style));
-        }
-        rows.push(ChartRow {
-            segments: x_label_segments,
-        });
-
-        if has_x_title {
-            let x_title_text = config.x_axis.title.as_deref().unwrap_or_default();
-            let x_title = center_text(x_title_text, plot_width);
-            let x_title_style = config.x_axis.title_style.unwrap_or(axis_style);
-            let mut x_title_segments: Vec<(String, Style)> = Vec::new();
-            if config.frame_visible {
-                x_title_segments.push(("│".to_string(), frame_style));
-            }
-            if config.y_axis_visible {
-                x_title_segments.push((footer_ylabel_pad, Style::new()));
-                x_title_segments.push((" ".repeat(y_axis_width), Style::new()));
-            }
-            x_title_segments.push((x_title, x_title_style));
-            x_title_segments.push((footer_legend_pad, Style::new()));
-            if config.frame_visible {
-                x_title_segments.push(("│".to_string(), frame_style));
-            }
-            rows.push(ChartRow {
-                segments: x_title_segments,
-            });
-        }
-    }
-
-    if config.frame_visible {
-        rows.push(ChartRow {
-            segments: vec![(format!("└{}┘", "─".repeat(content_width)), frame_style)],
-        });
-    }
-
-    rows
-}
-
-fn minimal_chart(
-    config: &ChartConfig,
-    width: usize,
-    frame_style: Style,
-    title_style: Style,
-) -> Vec<ChartRow> {
-    let mut rows = Vec::new();
-    if let Some(title) = &config.title {
-        rows.push(ChartRow {
-            segments: vec![(center_text(title, width), title_style)],
-        });
-    }
-    if config.frame_visible {
-        let inner = width.saturating_sub(2);
-        rows.push(ChartRow {
-            segments: vec![(format!("┌{}┐", "─".repeat(inner)), frame_style)],
-        });
-        rows.push(ChartRow {
-            segments: vec![(format!("│{}│", " ".repeat(inner)), frame_style)],
-        });
-        rows.push(ChartRow {
-            segments: vec![(format!("└{}┘", "─".repeat(inner)), frame_style)],
-        });
-    } else {
-        rows.push(ChartRow {
-            segments: vec![(" ".repeat(width), Style::new())],
-        });
-    }
-    rows
-}
-
-fn resolve_bounds<I>(values: I, manual: Option<(f64, f64)>) -> (f64, f64)
-where
-    I: Iterator<Item = f64>,
-{
-    if let Some((min, max)) = manual {
-        return normalize_bounds(min, max);
-    }
-
-    let mut min = f64::INFINITY;
-    let mut max = f64::NEG_INFINITY;
-    for value in values {
-        if !value.is_finite() {
-            continue;
-        }
-        min = min.min(value);
-        max = max.max(value);
-    }
-
-    if !min.is_finite() || !max.is_finite() {
-        return (0.0, 1.0);
-    }
-
-    normalize_bounds(min, max)
-}
-
-fn normalize_bounds(min: f64, max: f64) -> (f64, f64) {
-    if (max - min).abs() < f64::EPSILON {
-        let pad = if min.abs() < 1.0 {
-            1.0
-        } else {
-            min.abs() * 0.1
-        };
-        (min - pad, max + pad)
-    } else if min < max {
-        (min, max)
-    } else {
-        (max, min)
-    }
-}
-
-#[derive(Debug, Clone)]
-struct TickSpec {
-    values: Vec<f64>,
-    step: f64,
-}
-
-fn build_ticks(min: f64, max: f64, target: usize) -> TickSpec {
-    let span = (max - min).abs().max(f64::EPSILON);
-    let range = nice_number(span, false);
-    let raw_step = range / (target.max(2) as f64 - 1.0);
-    let step = nice_number(raw_step, true).max(f64::EPSILON);
-    let nice_min = (min / step).floor() * step;
-    let nice_max = (max / step).ceil() * step;
-
-    let mut values = Vec::new();
-    let mut value = nice_min;
-    let limit = nice_max + step * 0.5;
-    let mut guard = 0usize;
-    while value <= limit && guard < 128 {
-        values.push(value);
-        value += step;
-        guard = guard.saturating_add(1);
-    }
-
-    if values.is_empty() {
-        values.push(min);
-        values.push(max);
-    }
-
-    TickSpec { values, step }
-}
-
-/// TUI-aware tick generation: picks a nice step whose interval count
-/// divides `cell_count - 1` as evenly as possible, with 3-8 intervals
-/// and at least 2 rows per interval for readable spacing.
-fn build_tui_ticks(data_min: f64, data_max: f64, cell_count: usize) -> TickSpec {
-    let last = cell_count.saturating_sub(1).max(1);
-    let span = (data_max - data_min).abs().max(f64::EPSILON);
-    let log = span.log10().floor();
-
-    let mut candidates: Vec<(f64, f64, usize, usize)> = Vec::new();
-
-    for exp_off in -1..=1i32 {
-        let base = 10.0_f64.powf(log + f64::from(exp_off));
-        for &mult in &[1.0, 2.0, 2.5, 5.0] {
-            let step = base * mult;
-            if step <= 0.0 || !step.is_finite() {
-                continue;
-            }
-            let lo = (data_min / step).floor() * step;
-            let hi = (data_max / step).ceil() * step;
-            let n = ((hi - lo) / step + 0.5) as usize;
-            if (3..=8).contains(&n) && last / n >= 2 {
-                let rem = last % n;
-                candidates.push((step, lo, n, rem));
-            }
-        }
-    }
-
-    candidates.sort_by(|a, b| {
-        a.3.cmp(&b.3).then_with(|| {
-            let da = (a.2 as i32 - 5).unsigned_abs();
-            let db = (b.2 as i32 - 5).unsigned_abs();
-            da.cmp(&db)
-        })
-    });
-
-    if let Some(&(step, lo, n, _)) = candidates.first() {
-        let values: Vec<f64> = (0..=n).map(|i| lo + step * i as f64).collect();
-        return TickSpec { values, step };
-    }
-
-    build_ticks(data_min, data_max, 5)
-}
-
-fn nice_number(value: f64, round: bool) -> f64 {
-    if value <= 0.0 || !value.is_finite() {
-        return 1.0;
-    }
-    let exponent = value.log10().floor();
-    let power = 10.0_f64.powf(exponent);
-    let fraction = value / power;
-
-    let nice_fraction = if round {
-        if fraction < 1.5 {
-            1.0
-        } else if fraction < 3.0 {
-            2.0
-        } else if fraction < 7.0 {
-            5.0
-        } else {
-            10.0
-        }
-    } else if fraction <= 1.0 {
-        1.0
-    } else if fraction <= 2.0 {
-        2.0
-    } else if fraction <= 5.0 {
-        5.0
-    } else {
-        10.0
-    };
-
-    nice_fraction * power
-}
-
-fn format_number(value: f64, step: f64) -> String {
-    if !value.is_finite() {
-        return "0".to_string();
-    }
-    let abs_step = step.abs().max(f64::EPSILON);
-    let precision = if abs_step >= 1.0 {
-        0
-    } else {
-        (-abs_step.log10().floor() as i32 + 1).clamp(0, 6) as usize
-    };
-    format!("{value:.precision$}")
-}
-
-fn build_legend_items(datasets: &[Dataset]) -> Vec<(char, String, Color)> {
-    datasets
-        .iter()
-        .filter(|d| !d.name.is_empty())
-        .map(|d| {
-            let symbol = match d.graph_type {
-                GraphType::Line => '─',
-                GraphType::Area => '█',
-                GraphType::Scatter => marker_char(d.marker),
-                GraphType::Bar => '█',
-            };
-            (symbol, d.name.clone(), d.color)
-        })
-        .collect()
-}
-
-fn marker_char(marker: Marker) -> char {
-    match marker {
-        Marker::Braille => '⣿',
-        Marker::Dot => '•',
-        Marker::Block => '█',
-        Marker::HalfBlock => '▀',
-        Marker::Cross => '×',
-        Marker::Circle => '○',
-    }
-}
-
-struct GridSpec<'a> {
-    x_ticks: &'a [f64],
-    y_ticks: &'a [f64],
-    x_min: f64,
-    x_max: f64,
-    y_min: f64,
-    y_max: f64,
-}
-
-fn apply_grid(
-    config: &ChartConfig,
-    grid: GridSpec<'_>,
-    plot_chars: &mut [Vec<char>],
-    plot_styles: &mut [Vec<Style>],
-    grid_style: Style,
-) {
-    if !config.grid || plot_chars.is_empty() || plot_chars[0].is_empty() {
-        return;
-    }
-    let h = plot_chars.len();
-    let w = plot_chars[0].len();
-
-    for tick in grid.y_ticks {
-        let row = map_value_to_cell(*tick, grid.y_min, grid.y_max, h, true);
-        if row < h {
-            for col in 0..w {
-                if plot_chars[row][col] == ' ' {
-                    plot_chars[row][col] = '·';
-                    plot_styles[row][col] = grid_style;
-                }
-            }
-        }
-    }
-
-    for tick in grid.x_ticks {
-        let col = map_value_to_cell(*tick, grid.x_min, grid.x_max, w, false);
-        if col < w {
-            for row in 0..h {
-                if plot_chars[row][col] == ' ' {
-                    plot_chars[row][col] = '·';
-                    plot_styles[row][col] = grid_style;
-                }
-            }
-        }
-    }
-}
-
-fn draw_braille_dataset(
-    dataset: &Dataset,
-    x_min: f64,
-    x_max: f64,
-    y_min: f64,
-    y_max: f64,
-    plot_chars: &mut [Vec<char>],
-    plot_styles: &mut [Vec<Style>],
-) {
-    if dataset.data.is_empty() || plot_chars.is_empty() || plot_chars[0].is_empty() {
-        return;
-    }
-
-    let cols = plot_chars[0].len();
-    let rows = plot_chars.len();
-    let px_w = cols * 2;
-    let px_h = rows * 4;
-    let mut bits = vec![vec![0u32; cols]; rows];
-    let mut color_map = vec![vec![None::<Color>; cols]; rows];
-
-    let mut set_dot_colored = |px: usize, py: usize, color: Color| {
-        set_braille_dot(px, py, &mut bits, cols, rows);
-        let char_col = px / 2;
-        let char_row = py / 4;
-        if char_col < cols && char_row < rows {
-            color_map[char_row][char_col] = Some(color);
-        }
-    };
-
-    let points = dataset
-        .data
-        .iter()
-        .filter(|(x, y)| x.is_finite() && y.is_finite())
-        .map(|(x, y)| {
-            (
-                map_value_to_cell(*x, x_min, x_max, px_w, false),
-                map_value_to_cell(*y, y_min, y_max, px_h, true),
-                *y,
-            )
-        })
-        .collect::<Vec<_>>();
-
-    if points.is_empty() {
-        return;
-    }
-
-    if matches!(dataset.graph_type, GraphType::Line | GraphType::Area) {
-        let mut line_y_by_x = if matches!(dataset.graph_type, GraphType::Area) {
-            vec![None::<usize>; px_w]
-        } else {
-            Vec::new()
-        };
-        let mut line_color_by_x = if matches!(dataset.graph_type, GraphType::Area) {
-            vec![None::<Color>; px_w]
-        } else {
-            Vec::new()
-        };
-
-        for idx in 0..points.len().saturating_sub(1) {
-            let a = points[idx];
-            let b = points[idx + 1];
-            let seg_color = if let (Some(up), Some(down)) = (dataset.up_color, dataset.down_color) {
-                if b.2 > a.2 {
-                    up
-                } else {
-                    down
-                }
-            } else {
-                dataset.color
-            };
-
-            plot_bresenham(
-                a.0 as isize,
-                a.1 as isize,
-                b.0 as isize,
-                b.1 as isize,
-                |x, y| {
-                    if x < 0 || y < 0 {
-                        return;
-                    }
-                    let px = x as usize;
-                    let py = y as usize;
-                    set_dot_colored(px, py, seg_color);
-                    if matches!(dataset.graph_type, GraphType::Area) && px < px_w && py < px_h {
-                        line_y_by_x[px] = Some(match line_y_by_x[px] {
-                            Some(existing) => existing.min(py),
-                            None => py,
-                        });
-                        line_color_by_x[px] = Some(seg_color);
-                    }
-                },
-            );
-        }
-
-        if matches!(dataset.graph_type, GraphType::Area) {
-            for px in 0..px_w {
-                if let Some(line_y) = line_y_by_x[px] {
-                    let fill_color = line_color_by_x[px].unwrap_or(dataset.color);
-                    for py in line_y..px_h {
-                        set_dot_colored(px, py, fill_color);
-                    }
-                }
-            }
-        }
-    } else {
-        for (x, y, _) in &points {
-            set_dot_colored(*x, *y, dataset.color);
-        }
-    }
-
-    for row in 0..rows {
-        for col in 0..cols {
-            if bits[row][col] != 0 {
-                let ch = char::from_u32(BRAILLE_BASE + bits[row][col]).unwrap_or(' ');
-                plot_chars[row][col] = ch;
-                let color = color_map[row][col].unwrap_or(dataset.color);
-                plot_styles[row][col] = Style::new().fg(color);
-            }
-        }
-    }
-
-    if !matches!(dataset.marker, Marker::Braille) {
-        let m = marker_char(dataset.marker);
-        for (x, y) in dataset
-            .data
-            .iter()
-            .filter(|(x, y)| x.is_finite() && y.is_finite())
-        {
-            let col = map_value_to_cell(*x, x_min, x_max, cols, false);
-            let row = map_value_to_cell(*y, y_min, y_max, rows, true);
-            if row < rows && col < cols {
-                plot_chars[row][col] = m;
-                plot_styles[row][col] = Style::new().fg(dataset.color);
-            }
-        }
-    }
-}
-
-fn draw_bar_dataset(
-    dataset: &Dataset,
-    _x_min: f64,
-    _x_max: f64,
-    y_min: f64,
-    y_max: f64,
-    plot_chars: &mut [Vec<char>],
-    plot_styles: &mut [Vec<Style>],
-) {
-    if dataset.data.is_empty() || plot_chars.is_empty() || plot_chars[0].is_empty() {
-        return;
-    }
-
-    let rows = plot_chars.len();
-    let cols = plot_chars[0].len();
-    let n = dataset.data.len();
-    let slot_width = cols as f64 / n as f64;
-    let zero_row = map_value_to_cell(0.0, y_min, y_max, rows, true);
-
-    for (index, (_, value)) in dataset.data.iter().enumerate() {
-        if !value.is_finite() {
-            continue;
-        }
-
-        let start_f = index as f64 * slot_width;
-        let bar_width_f = (slot_width * 0.75).max(1.0);
-        let full_w = bar_width_f.floor() as usize;
-        let frac_w = ((bar_width_f - full_w as f64) * 8.0).round() as usize;
-
-        let x_start = start_f.floor() as usize;
-        let x_end = (x_start + full_w).min(cols.saturating_sub(1));
-        let frac_col = (x_end + 1).min(cols.saturating_sub(1));
-
-        let value_row = map_value_to_cell(*value, y_min, y_max, rows, true);
-        let (top, bottom) = if value_row <= zero_row {
-            (value_row, zero_row)
-        } else {
-            (zero_row, value_row)
-        };
-
-        for row in top..=bottom.min(rows.saturating_sub(1)) {
-            for col in x_start..=x_end {
-                if col < cols {
-                    plot_chars[row][col] = '█';
-                    plot_styles[row][col] = Style::new().fg(dataset.color);
-                }
-            }
-            if frac_w > 0 && frac_col < cols {
-                plot_chars[row][frac_col] = BLOCK_FRACTIONS[frac_w.min(8)];
-                plot_styles[row][frac_col] = Style::new().fg(dataset.color);
-            }
-        }
-    }
-}
-
-fn overlay_legend_on_plot(
-    position: LegendPosition,
-    items: &[(char, String, Color)],
-    plot_chars: &mut [Vec<char>],
-    plot_styles: &mut [Vec<Style>],
-    axis_style: Style,
-) {
-    if plot_chars.is_empty() || plot_chars[0].is_empty() || items.is_empty() {
-        return;
-    }
-
-    let rows = plot_chars.len();
-    let cols = plot_chars[0].len();
-    let start_row = match position {
-        LegendPosition::TopLeft => 0,
-        LegendPosition::BottomLeft => rows.saturating_sub(items.len()),
-        _ => 0,
-    };
-
-    for (i, (symbol, name, color)) in items.iter().enumerate() {
-        let row = start_row + i;
-        if row >= rows {
-            break;
-        }
-        let legend_text = format!("{symbol} {name}");
-        for (col, ch) in legend_text.chars().enumerate() {
-            if col >= cols {
-                break;
-            }
-            plot_chars[row][col] = ch;
-            plot_styles[row][col] = if col == 0 {
-                Style::new().fg(*color)
-            } else {
-                axis_style
-            };
-        }
-    }
-}
-
-fn build_y_tick_row_map(
-    ticks: &[f64],
-    labels: Option<&[String]>,
-    y_min: f64,
-    y_max: f64,
-    plot_height: usize,
-) -> Vec<(usize, String)> {
-    let step = if ticks.len() > 1 {
-        (ticks[1] - ticks[0]).abs()
-    } else {
-        1.0
-    };
-    ticks
-        .iter()
-        .enumerate()
-        .map(|(idx, v)| {
-            let label = labels
-                .and_then(|manual| manual.get(idx).cloned())
-                .unwrap_or_else(|| format_number(*v, step));
-            (
-                map_value_to_cell(*v, y_min, y_max, plot_height, true),
-                label,
-            )
-        })
-        .collect()
-}
-
-fn build_x_tick_col_map(
-    ticks: &[f64],
-    labels: Option<&[String]>,
-    labels_match_manual_ticks: bool,
-    x_min: f64,
-    x_max: f64,
-    plot_width: usize,
-) -> Vec<(usize, String)> {
-    if let Some(labels) = labels {
-        if labels.is_empty() {
-            return Vec::new();
-        }
-        if labels_match_manual_ticks {
-            return ticks
-                .iter()
-                .zip(labels.iter())
-                .map(|(tick, label)| {
-                    (
-                        map_value_to_cell(*tick, x_min, x_max, plot_width, false),
-                        label.clone(),
-                    )
-                })
-                .collect();
-        }
-        let denom = labels.len().saturating_sub(1).max(1);
-        return labels
-            .iter()
-            .enumerate()
-            .map(|(i, label)| {
-                let col = (i * plot_width.saturating_sub(1)) / denom;
-                (col, label.clone())
-            })
-            .collect();
-    }
-
-    let step = if ticks.len() > 1 {
-        (ticks[1] - ticks[0]).abs()
-    } else {
-        1.0
-    };
-    ticks
-        .iter()
-        .map(|v| {
-            (
-                map_value_to_cell(*v, x_min, x_max, plot_width, false),
-                format_number(*v, step),
-            )
-        })
-        .collect()
-}
-
-fn map_value_to_cell(value: f64, min: f64, max: f64, size: usize, invert: bool) -> usize {
-    if size == 0 {
-        return 0;
-    }
-    let span = (max - min).abs().max(f64::EPSILON);
-    let mut t = ((value - min) / span).clamp(0.0, 1.0);
-    if invert {
-        t = 1.0 - t;
-    }
-    (t * (size.saturating_sub(1)) as f64).round() as usize
-}
-
-fn set_braille_dot(px: usize, py: usize, bits: &mut [Vec<u32>], cols: usize, rows: usize) {
-    if cols == 0 || rows == 0 {
-        return;
-    }
-    let char_col = px / 2;
-    let char_row = py / 4;
-    if char_col >= cols || char_row >= rows {
-        return;
-    }
-    let sub_col = px % 2;
-    let sub_row = py % 4;
-    bits[char_row][char_col] |= if sub_col == 0 {
-        BRAILLE_LEFT_BITS[sub_row]
-    } else {
-        BRAILLE_RIGHT_BITS[sub_row]
-    };
-}
-
-fn plot_bresenham(x0: isize, y0: isize, x1: isize, y1: isize, mut plot: impl FnMut(isize, isize)) {
-    let mut x = x0;
-    let mut y = y0;
-    let dx = (x1 - x0).abs();
-    let sx = if x0 < x1 { 1 } else { -1 };
-    let dy = -(y1 - y0).abs();
-    let sy = if y0 < y1 { 1 } else { -1 };
-    let mut err = dx + dy;
-
-    loop {
-        plot(x, y);
-        if x == x1 && y == y1 {
-            break;
-        }
-        let e2 = 2 * err;
-        if e2 >= dy {
-            err += dy;
-            x += sx;
-        }
-        if e2 <= dx {
-            err += dx;
-            y += sy;
-        }
-    }
-}
-
-fn center_text(text: &str, width: usize) -> String {
-    let text_width = UnicodeWidthStr::width(text);
-    if text_width >= width {
-        return text.chars().take(width).collect();
-    }
-    let left = (width - text_width) / 2;
-    let right = width - text_width - left;
-    format!("{}{}{}", " ".repeat(left), text, " ".repeat(right))
-}
-
-fn sturges_bin_count(n: usize) -> usize {
-    if n <= 1 {
-        return 1;
-    }
-    (1.0 + (n as f64).log2()).ceil() as usize
 }
